Moved all Files into Subfolders & Made some other adjustments 1) Moved all files into subfolders in order to make things more neater and concise.
2) Changed the speed of the user plane in order to promote a more exciting playtime

3) Added hash set to keep track of active keys and prevent weird results happening from pressing two keys at once (i.e. Holding space to fire while moving would make the fire stutter and feel unrewarding/make the controls feel unresponsive). With hash set this was solved.

4) Added a temporary 3rd level with adjusted difficulties for testing.
