+++ src/main/java/com/example/demo/actors/FighterPlane.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.actors;
 
 public abstract class FighterPlane extends ActiveActorDestructible {
 
@@ -10,7 +10,7 @@ public abstract class FighterPlane extends ActiveActorDestructible {
 	}
 
 	public abstract ActiveActorDestructible fireProjectile();
-	
+
 	@Override
 	public void takeDamage() {
 		health--;
@@ -34,5 +34,4 @@ public abstract class FighterPlane extends ActiveActorDestructible {
 	public int getHealth() {
 		return health;
 	}
-		
 }
